+++ Orginal/Source/CommandCharaSelect.cpp
@@ -17,21 +17,23 @@ CommandCharaSelect::CommandCharaSelect(Actor::Type characterType)
 
 void CommandCharaSelect::Update(const BattleActorManager* bam, CommandBase* cmdBase)
 {
+	// キャラタイプによってStateを変える
 	if (mCharaType == Actor::PLAYER)     BattleState::GetInstance().SetState(BattleState::State::PLAYER_SELECT);
 	else if (mCharaType == Actor::ENEMY) BattleState::GetInstance().SetState(BattleState::State::ENEMY_SELECT);
 
+	// 生きてる同じタイプのアクターから選択する
 	const std::vector<int>& ids = bam->GetAliveActorIDs(mCharaType);
-
-	size_t max = ids.size() - 1;
-	if (Input::GetButtonTrigger(0, Input::BUTTON::RIGHT)) mCharaIndex = Math::Min(mCharaIndex + 1, static_cast<int>(max));
+	int max = static_cast<int>(ids.size() - 1);
+	if (Input::GetButtonTrigger(0, Input::BUTTON::RIGHT)) mCharaIndex = Math::Min(mCharaIndex + 1, max);
 	if (Input::GetButtonTrigger(0, Input::BUTTON::LEFT))  mCharaIndex = Math::Max(mCharaIndex - 1, 0);
-
 	mTargetActor = bam->GetActor(ids[mCharaIndex]);
 
+	// キャラ選択したら
 	if (Input::GetButtonTrigger(0, Input::BUTTON::A))
 	{
 		cmdBase->SetTargetObjID(ids[mCharaIndex]);
 
+		// アイテムが登録されているかどうかで分岐
 		if (cmdBase->GetItemParam()) cmdBase->SetBehaviour(CommandBase::Behaviour::USE_ITEM);
 		else						 cmdBase->SetBehaviour(CommandBase::Behaviour::ATTACK);
 	}
